# Shared/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "BlazorFaceRecog.Server",
    "version": "1.0"
  },
  "paths": {
    "/Faces": {
      "get": {
        "tags": [
          "FaceEndpoints"
        ],
        "operationId": "GetSavedFaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedFaceModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Faces/Train": {
      "post": {
        "tags": [
          "FaceEndpoints"
        ],
        "operationId": "TrainFaces",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TrainFaceModel"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/Faces/Detect": {
      "post": {
        "tags": [
          "FaceEndpoints"
        ],
        "operationId": "DetectFaces",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file",
                  "id"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                },
                "id": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SavedFaceModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "thumbnail": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TrainFaceModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}